Use backticks for code formatting and add links
Congratulations on the preprint! Very interesting work. Homology detection is a difficult and unsolved problem and I appreciated your take on it.

This PR changes some formatting in the readme to make it easier to copy-paste the code output, and read what is code vs what is text. Additionally, this changes some of the bold text that looked like headings, into Header 3 values with three levels via `###`. Removed double underscore bolding for headings as that is implied by the header tag. Also, paper and server links were added in a few places where specified. Table of contents was generated by [this tool][https://ecotrust-canada.github.io/markdown-toc/)